Add a minimal app framework
This looks a lot like Context and State from Tetra, but the key difference is that App/EventHandler *aren't* baked into the rest of the framework - they're just a convenient way of getting a game loop up and running. Creating a game specific event loop should be as simple as copying and pasting app.rs into the project.