Simplify batching API
There's two bits to this commit:

First, I've shuffled stuff around with the graphics internals so that Batcher can access the graphics context without it being passed in - it's not possible to use a different context than the one it was created with, so having it as a parameter was just noise.

Second, I've removed the 'clear' function and made that part of the RenderPass API instead (with functions to expose this on Batcher) - this is more in line with how modern graphics APIs (and wrappers around them like Sokol and WGPU) handle clearing, so should futureproof me a bit.