Revert change to `clear` API
The new API ended up being awkward to use when you're doing stuff without a batcher (e.g. UI library integrations), so it needs a rethink. I'd like to get to a point where you can just call `clear` on the target itself, but that doesn't play nicely with `Window`...